// src/types/courseManagement.ts
import { TAcademicSemester } from "./academicManagement.type"

export type TSemester={
    _id:string,
    academicSemester:TAcademicSemester
    status: string,
    startDate:string,
    endDate:string,
    minCredit: number;
    maxCredit: number;
    createdAt: string;
    updatedAt: string;
}

export type TCourse = {
    fullName: any;
    _id: string;
    title: string;
    prefix: string;
    code: number;
    credits: number;
    preRequisiteCourses: { course: string | null; isDeleted: boolean }[];
    isDeleted: boolean;
  };